Sorted out the up/down tool organization.
1. Needed to refactor the tool rendering functions in the
   catalog to allow visible layers to add the up/down tools
   if not in the catalog by default.
2. Moved some of the ZOrdering functions to util so they
   can be used in the catalog(s) and in the tools.
